exomeseq vcf sqlite3 database generation
This needs to be tested further. The lines of code that generate the database have been added to each rule. In short, the rule now generates a sqlite commands file which includes the table declaration commands for headers and importing commands. The temp files generated are generated after some processing and used for importing into sqlite3. Some of the processing includes removing special characters from the headers. Additionally, since I noticed some columns are duplicates, or at least have duplicate headers.. I am using R to remove some of the said duplicated rows. Another option would be to change the R script that I am using (dedupcols.R) and comment out the deduplication step (line 2), and change check.names to True in the table import step. I believe this will force the column names to be appended with a unique number if there are duplicates, and thus could solve the issue in a workaround way without removing columns. I also finally remove the header row in the temp file before importing, after the header information is already sent to the sqlite commands script. The file generated is called vcf.db. NOTE. it may be best to add a line before all of this to remove the file vcf.db if it already exists. This is because if the pipeline is re-run again sqlite3 will throw an error since the file already exists. If there are any bugs with these steps then it should not be a problem as Snakemake will still find the output files frmo the previous lines of code. If there are any issues please contact me at nikhilbg96@gmail.com. You will also find the VariantFilteringTool repository updated with a version that is compatible with the vcf.db file created here.